Add InitialTestRun trait
